Initial Commit
This adds the ontology map creation.

There are several areas of improvement needed:
Origin position
Coloring of nodes to be fixed

This branch is unstable, it the ontology map does not instantiate its objects across the network and is test environment code right now.

I'll fix this code tomorrow and re-enable the disabled interactables.